Refactor share sheet
